+++ test/e2e/runs.test.js
@@ -240,7 +240,17 @@ module.exports = () => {
 
             const { data } = response.body;
             // 48 because one run is added in start run use case with default quality which is test, and one is updated to quality test
-            expect(data.length).to.equal(48);
+            expect(data.length).to.equal(47);
+        });
+        it('should filter run on their trigger value', async () => {
+            const response = await request(server)
+                .get('/api/runs?filter[triggerValues]=OFF,CTP');
+            expect(response.status).to.equal(200);
+
+            expect(response).to.satisfyApiSpec;
+
+            const { data } = response.body;
+            expect(data.length).to.equal(12);
         });
         it('should return 400 if "runQuality" is invalid', async () => {
             const response = await request(server)
@@ -277,6 +287,18 @@ module.exports = () => {
             expect(data).to.have.lengthOf(52);
         });
 
+        it('should successfully filter on lhcPeriod', async () => {
+            const response =
+                await request(server).get('/api/runs?filter[lhcPeriods]=lhc22_b');
+
+            expect(response.status).to.equal(200);
+            expect(response).to.satisfyApiSpec;
+
+            const { data } = response.body;
+            expect(data).to.be.an('array');
+            expect(data).to.lengthOf.above(1);
+        });
+
         it('should return 400 if the FLP number filter is invalid', async () => {
             const response = await request(server).get('/api/runs?filter[nFlps][operator]=invalid&filter[nFlps][limit]=10');
 
@@ -288,7 +310,7 @@ module.exports = () => {
             expect(error.detail).to.equal('"query.filter.nFlps.operator" must be one of [<, <=, =, >=, >]');
         });
 
-        it('should successfully filter on detectors number', async () => {
+        it('should successfully filter on FLPs number', async () => {
             const response =
                 await request(server).get('/api/runs?filter[nFlps][operator]=<=&filter[nFlps][limit]=10');
 
@@ -298,6 +320,30 @@ module.exports = () => {
             const { data } = response.body;
             expect(data).to.be.an('array');
             // 7 instead of 5 because 2 runs have been created with 10 as nFlps
+            expect(data).to.have.lengthOf(7);
+        });
+
+        it('should return 400 if the EPN number filter is invalid', async () => {
+            const response = await request(server).get('/api/runs?filter[nEpns][operator]=invalid&filter[nEpns][limit]=10');
+
+            expect(response.status).to.equal(400);
+            expect(response).to.satisfyApiSpec;
+
+            const { errors: [error] } = response.body;
+            expect(error.title).to.equal('Invalid Attribute');
+            expect(error.detail).to.equal('"query.filter.nEpns.operator" must be one of [<, <=, =, >=, >]');
+        });
+
+        it('should successfully filter on EPNs number', async () => {
+            const response =
+                await request(server).get('/api/runs?filter[nEpns][operator]=<=&filter[nEpns][limit]=10');
+
+            expect(response.status).to.equal(200);
+            expect(response).to.satisfyApiSpec;
+
+            const { data } = response.body;
+            expect(data).to.be.an('array');
+            // 7 instead of 5 because 2 runs have been created with 10 as nEpns
             expect(data).to.have.lengthOf(7);
         });
     });
@@ -518,6 +564,7 @@ module.exports = () => {
                         done(err);
                         return;
                     }
+                    expect(res.body.data.triggerValue).to.equal('OFF');
                     expect(res.body.data).to.be.an('object');
                     expect(res.body.data.id).to.equal(109);
 
@@ -557,6 +604,7 @@ module.exports = () => {
             expect(response.body.errors[0].detail).to.equal('A run already exists with run number 109');
         });
     });
+
     describe('PUT /api/runs/:runId', () => {
         it('should return 200 in all other cases', (done) => {
             request(server)
@@ -600,11 +648,11 @@ module.exports = () => {
                 .put('/api/runs/1')
                 .expect(201)
                 .send({
-                    runQuality: 'test',
+                    runQuality: 'good',
                 });
             expect(body.data).to.be.an('object');
             expect(body.data.id).to.equal(1);
-            expect(body.data.runQuality).to.equal('test');
+            expect(body.data.runQuality).to.equal('good');
         });
 
         it('should return an error due to invalid runQuality value', async () => {
@@ -615,7 +663,6 @@ module.exports = () => {
                     runQuality: 'wrong',
                 });
             expect(body.errors).to.be.an('array');
-            // eslint-disable-next-line max-len
             expect(body.errors[0].detail).to.equal('"body.runQuality" must be one of [good, bad, test]');
         });
 
@@ -637,12 +684,25 @@ module.exports = () => {
             expect(body.errors[0].detail).to.equal('"body.eorReasons[0].runId" must be a number');
         });
 
-        it('should successfully add eorReasons to run', async () => {
+        it('should successfully return the updated run entity with new runQuality value', async () => {
+            const { body } = await request(server)
+                .put('/api/runs/106')
+                .expect(201)
+                .send({
+                    runQuality: 'good',
+                });
+            expect(body.data).to.be.an('object');
+            expect(body.data.id).to.equal(106);
+            expect(body.data.runQuality).to.equal('good');
+        });
+
+        it('should successfully add eorReasons to run and check runQuality did not change', async () => {
             const currentRun = await request(server)
                 .get('/api/runs/106')
                 .expect(200);
             expect(currentRun.body.data).to.be.an('object');
             expect(currentRun.body.data.id).to.equal(106);
+            expect(currentRun.body.data.runQuality).to.equal('good');
             expect(currentRun.body.data.eorReasons).to.have.lengthOf(0);
 
             const { body } = await request(server)
@@ -662,6 +722,7 @@ module.exports = () => {
             expect(body.data.id).to.equal(106);
             expect(body.data.eorReasons).to.have.lengthOf(1);
             expect(body.data.eorReasons[0].description).to.equal('Some');
+            expect(body.data.runQuality).to.equal('good');
         });
     });
 
@@ -722,9 +783,9 @@ module.exports = () => {
                 });
         });
     });
-    describe('PATCH api/runs/:runId', () => {
+    describe('PATCH api/runs/:runNumber', () => {
         const dateValue = new Date('1-1-2021').setHours(0, 0, 0, 0);
-        it('should return 400 when runId is wrong', (done) => {
+        it('should return 400 when runNumber is wrong', (done) => {
             request(server)
                 .patch('/api/runs/9999999999')
                 .send({
@@ -745,13 +806,14 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should return 201 in all other cases', (done) => {
+        it('should successfully update a run by its RunNumber with partial information', (done) => {
             request(server)
                 .patch('/api/runs/1')
                 .send({
                     timeO2End: dateValue,
                     timeTrgEnd: dateValue,
-                    runQuality: 'test',
+                    runQuality: 'good',
+                    lhcPeriod: 'lhc22_b',
                 })
                 .expect(201)
                 .end((err, res) => {
@@ -763,37 +825,34 @@ module.exports = () => {
                     expect(res.body.data.id).to.equal(1);
                     expect(res.body.data.timeO2End).to.equal(dateValue);
                     expect(res.body.data.timeTrgEnd).to.equal(dateValue);
-                    expect(res.body.data.runQuality).to.equal('test');
+                    expect(res.body.data.runQuality).to.equal('good');
+                    expect(res.body.data.lhcPeriod).to.equal('lhc22_b');
                     done();
                 });
         });
-        it('should be able to update with ', async () => {
+
+        it('should successfully update a run by its RunNumber with partial information and keep previous updated values the same', async () => {
             const { body } = await request(server)
-                .patch('/api/runs/80')
+                .patch('/api/runs/1')
                 .expect(201)
                 .send({
-                    runQuality: 'test',
-                    timeO2End: dateValue,
                     timeO2Start: dateValue,
                     timeTrgStart: dateValue,
-                    timeTrgEnd: dateValue,
                     pdpConfigOption: 'Repository hash',
-                    trgGlobalRunEnabled: true,
-                    trgEnabled: false,
                     pdpTopologyDescriptionLibraryFile: 'production/production.desc',
                     tfbDdMode: 'processing',
+                    lhcPeriod: 'lhc22_b',
+                    triggerValue: 'LTU',
                 });
             expect(body.data).to.be.an('object');
-            expect(body.data.timeO2End).to.equal(dateValue);
-            expect(body.data.timeO2Start).to.equal(dateValue);
-            expect(body.data.timeTrgStart).to.equal(dateValue);
-            expect(body.data.timeTrgEnd).to.equal(dateValue);
-            expect(body.data.runQuality).to.equal('test');
+            expect(body.data.timeO2End).to.equal(dateValue); // Values not passed should remain the same
+            expect(body.data.timeO2Start).to.equal(dateValue); // Values not passed should remain the same
+            expect(body.data.runQuality).to.equal('good'); // Values not passed should remain the same
             expect(body.data.pdpConfigOption).to.equal('Repository hash');
-            expect(body.data.trgGlobalRunEnabled).to.equal(true);
-            expect(body.data.trgEnabled).to.equal(false);
             expect(body.data.pdpTopologyDescriptionLibraryFile).to.equal('production/production.desc');
             expect(body.data.tfbDdMode).to.equal('processing');
+            expect(body.data.lhcPeriod).to.equal('lhc22_b');
+            expect(body.data.triggerValue).to.equal('LTU');
         });
     });
 };