+++ test/public/runs/overview.test.js
@@ -295,7 +295,7 @@ module.exports = () => {
 
         await checkColumnBalloon(page, 1, 2);
         await checkColumnBalloon(page, 1, 3);
-        await checkColumnBalloon(page, 1, 14);
+        await checkColumnBalloon(page, 1, 17);
     });
 
     it('Should display balloon if the text overflows', async () => {
@@ -516,6 +516,50 @@ module.exports = () => {
         await checkTableRunQualities(table, ['bad']);
     });
 
+    it('Should successfully filter runs by their trigger value', async () => {
+        await page.goto(`${url}?page=run-overview`, { waitUntil: 'networkidle0' });
+        const filterInputSelectorPrefix = '#triggerValueCheckbox';
+        const offFilterSelector = `${filterInputSelectorPrefix}OFF`;
+        const ltuFilterSelector = `${filterInputSelectorPrefix}LTU`;
+
+        /**
+         * Checks that all the rows of the given table have a valid trigger value
+         *
+         * @param {{evaluate: function}[]} rows the list of rows
+         * @param {string[]} authorizedRunQualities  the list of valid run qualities
+         * @return {void}
+         */
+        const checkTableRunQualities = async (rows, authorizedRunQualities) => {
+            for (const row of rows) {
+                expect(await row.evaluate((rowItem) => {
+                    const rowId = rowItem.id;
+                    return document.querySelector(`#${rowId}-triggerValue-text`).innerText;
+                })).to.be.oneOf(authorizedRunQualities);
+            }
+        };
+
+        // Open filter toggle
+        await pressElement(page, '#openRunFilterToggle');
+        await page.waitForTimeout(200);
+
+        await page.$eval(offFilterSelector, (element) => element.click());
+        await page.waitForTimeout(300);
+        table = await page.$$('tbody tr');
+        expect(table.length).to.equal(6);
+        await checkTableRunQualities(table, ['OFF']);
+
+        await page.$eval(ltuFilterSelector, (element) => element.click());
+        await page.waitForTimeout(300);
+        table = await page.$$('tbody tr');
+        await checkTableRunQualities(table, ['OFF', 'LTU']);
+
+        await page.$eval(ltuFilterSelector, (element) => element.click());
+        await page.waitForTimeout(300);
+        table = await page.$$('tbody tr');
+        expect(table.length).to.equal(6);
+        await checkTableRunQualities(table, ['OFF']);
+    });
+
     it('should successfully filter on a list of run ids and inform the user about it', async () => {
         await page.reload();
         await page.waitForTimeout(200);
@@ -617,6 +661,35 @@ module.exports = () => {
             return document.querySelector(`#${rowId}-nFlps-text`)?.innerText;
         }));
         expect(nFlpsList.every((nFlps) => parseInt(nFlps, 10) <= '10')).to.be.true;
+    });
+
+    it('should successfully filter on nEpns', async () => {
+        await page.goto(`${url}?page=run-overview`, { waitUntil: 'networkidle0' });
+        page.waitForTimeout(100);
+
+        await pressElement(page, '#openRunFilterToggle');
+        await page.waitForTimeout(200);
+
+        const nEpnsOperatorSelector = '#nEpns-operator';
+        const nEpnsOperator = await page.$(nEpnsOperatorSelector) || null;
+        expect(nEpnsOperator).to.not.be.null;
+        expect(await nEpnsOperator.evaluate((element) => element.value)).to.equal('=');
+
+        const nEpnsLimitSelector = '#nEpns-limit';
+        const nEpnsLimit = await page.$(nEpnsLimitSelector) || null;
+        expect(nEpnsLimit).to.not.be.null;
+
+        await nEpnsLimit.focus();
+        await page.keyboard.type('10');
+        await page.waitForTimeout(300);
+        await page.select(nEpnsOperatorSelector, '<=');
+        await page.waitForTimeout(300);
+
+        const nEpnsList = await page.evaluate(() => Array.from(document.querySelectorAll('tbody tr')).map((row) => {
+            const rowId = row.id;
+            return document.querySelector(`#${rowId}-nEpns-text`)?.innerText;
+        }));
+        expect(nEpnsList.every((nEpns) => parseInt(nEpns, 10) <= '10')).to.be.true;
     });
 
     const EXPORT_RUNS_TRIGGER_SELECTOR = '#export-runs-trigger';